NewGame sceen done (needs Map object)
