+++ src/screens/__init__.py
@@ -7,6 +7,7 @@ MIT license, see LICENSE.md for details.
 from .Base import Base
 from .Credits import Credits
 from .Journey import Journey
+from .NewGame import NewGame
 from .Presents import Presents
 from .Settings import Settings
 from .Title import Title
@@ -15,6 +16,7 @@ __all__ = [
     'Base',
     'Credits',
     'Journey',
+    'NewGame',
     'Presents',
     'Settings',
     'Title',
